pyproject.toml
[tool.poetry]
name = "clingo"
version = "0.1.0"
description = "Learn languages by watching video clips"
authors = ["Ace"]
license = "MIT"
readme = "README.md"
packages = [{include = "clingo"}]

[tool.poetry.dependencies]
python = "^3.11"
youtube-dl = {git = "https://github.com/ytdl-org/ytdl-nightly.git", rev = "2024.02.03"}
playwright = "^1.41.2"
moviepy = {git = "https://github.com/Zulko/moviepy.git"}
vtt-to-srt3 = "^0.2.0.1"
pydeeplx = "^1.0.7"
django = "^5.0.3"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"



[tool.poetry.group.dev.dependencies]
pytest-playwright = "^0.4.4"
pre-commit = "^3.6.2"
ruff = "^0.3.2"
mypy = "^1.9.0"
django-stubs = "^4.2.7"
django-stubs-ext = "^4.2.7"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
pytest-django-queries = "^1.2.0"
pytest-factoryboy = "^2.7.0"
pytest-html = "^4.1.1"
pytest-mock = "^3.12.0"

[tool.ruff]
lint.extend-select = ["D", "I", "ISC002", "PIE", "PT", "T20", "UP"]
lint.ignore = ["D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", "D202", "D203", "D206", "D213", "D407", "PT019"]
target-version = "py311"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.isort]
known-first-party = ["clingo"]

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "clingo.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "clingo.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
